refactor(components.api): move to webUniverseComponents
ref: OM-148
