+++ client/app/hosting/order/hosting-order.service.js
@@ -3,17 +3,17 @@ angular.module('services').service(
   class HostingOrder {
     /**
      * Constructor
-     * @param Api
+     * @param WucApi
      * @param constants
      */
-    constructor(Api, constants) {
-      this.Api = Api;
+    constructor(WucApi, constants) {
+      this.WucApi = WucApi;
       this.constants = constants;
       this.proxyPass = `${constants.swsProxyRootPath}order/hosting/web`;
     }
 
     getDurations(domain, offer, dnsZone) {
-      return this.Api.get(`${this.proxyPass}/new`, {
+      return this.WucApi.get(`${this.proxyPass}/new`, {
         params: {
           dnsZone,
           domain,
@@ -33,7 +33,7 @@ angular.module('services').service(
       if (module) {
         parameters.module = module;
       }
-      return this.Api.get(`${this.proxyPass}/new/${duration}`, {
+      return this.WucApi.get(`${this.proxyPass}/new/${duration}`, {
         params: parameters,
       });
     }
@@ -49,13 +49,13 @@ angular.module('services').service(
       if (module) {
         parameters.module = module;
       }
-      return this.Api.post(`${this.proxyPass}/new/${duration}`, {
+      return this.WucApi.post(`${this.proxyPass}/new/${duration}`, {
         data: parameters,
       });
     }
 
     getModels() {
-      return this.Api.get(`${this.constants.swsProxyRootPath}order.json`).then(data => data.models);
+      return this.WucApi.get(`${this.constants.swsProxyRootPath}order.json`).then(data => data.models);
     }
   },
 );
