+++ client/app/webUniverseComponents/api/index.js
@@ -1,0 +1,11 @@
+import angular from 'angular';
+
+import WucApi from './Api';
+
+const moduleName = 'wucApi';
+
+angular
+  .module(moduleName, [])
+  .service('WucApi', WucApi);
+
+export default moduleName;
